Generate the file array buffer only once
