+++ packages/ui/src/common/utils/file.ts
@@ -1,6 +1,3 @@
 import { memoize } from 'lodash'
 
-export const asUint8Array = memoize(async (file: File) => {
-  const arrayBuffer = await file.arrayBuffer()
-  return new Uint8Array(arrayBuffer)
-})
+export const asArrayBuffer = memoize((file?: File) => file?.arrayBuffer() ?? new ArrayBuffer(0))
